+++ apps/app/src/client/services/yjs.ts
@@ -1,8 +1,4 @@
-import { useEffect, useMemo } from 'react';
-
-import { GlobalCodeMirrorEditorKey } from '@growi/editor';
-import { useCodeMirrorEditorIsolated } from '@growi/editor/dist/client/stores/codemirror-editor';
-import { useTranslation } from 'next-i18next';
+import { useMemo } from 'react';
 
 import { useYjsMaxBodyLength } from '~/stores-universal/context';
 import { EditorMode, useEditorMode } from '~/stores-universal/ui';
@@ -27,26 +23,4 @@ export const useIsYjsEnabled = (): boolean | undefined => {
   }, [shouldRecalculate]);
 
   return isYjsEnabled;
-};
-
-export const useNonYjsModeEffect = (): void => {
-  const { t } = useTranslation();
-  const { data: currentPage } = useSWRxCurrentPage();
-  const { data: yjsMaxBodyLength } = useYjsMaxBodyLength();
-  const { data: codeMirrorEditor } = useCodeMirrorEditorIsolated(GlobalCodeMirrorEditorKey.MAIN);
-  const isYjsEnabled = useIsYjsEnabled();
-
-  // initDoc() is not available unless codeMirrorEditor.view is true
-  const shouldRecalculate = isYjsEnabled === false && codeMirrorEditor != null && codeMirrorEditor.view != null;
-
-  useEffect(() => {
-    if (shouldRecalculate) {
-      codeMirrorEditor.initDoc(currentPage?.revision?.body);
-
-      // eslint-disable-next-line no-alert
-      window.alert(t('non-yjs-alert', { yjsMaxBodyLength }));
-    }
-
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [shouldRecalculate]);
 };
